Move getApiUlrs to lib/api
